Solution for Day 19
A lot of code as been reused from Day 16. As an optimization, instruction names are remapped to opcodes when read to avoid string comparisons during execution.
Part 2 might not be generic as I haven't checked other data sets. I have basically optimized what the input program does to be as efficient as possible.