Solutions for Day 20 and 21
Day 20 is a generic solution for any kind of regex expression. It works in three passes: Bounds finding, connection creations and distance computation via a fill algorithm.
Day 21 expects eqrr instruction to be used for comparison (gtrr or any immediate comparison do not make sense for the puzzle). Although untested on other input sets, the solution is general.